feat: Add BulkTriggerSidebar component and integrate it into DeployImageContent for improved image selection handling
